Moved breakpoint toggle into HoverButton
